% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosineSimilarity.R
\name{cosineSimilarity}
\alias{cosineSimilarity}
\title{Calculate the cosine similarity between two numeric vectors}
\usage{
cosineSimilarity(d1, d2)
}
\arguments{
\item{d1}{First numeric vector}

\item{d2}{Second numeric vector (must be the same length as d1)}
}
\value{
A numeric vector of cosine similarities.
}
\description{
Calculate the cosine similarity between two numeric vectors
}
\details{
This function calculates the cosine similarity between a pair of vectors.

Cosine similarities are a measure of similarity between two series of numbers. They are mathematically defined as the cosine of the angle between both vectors.
Cosine similarities are bound between -1 and 1, with −1 indicating exactly opposite directions and 1 indicating identical directions. Values of 0 indicate the vectors are orthogonal.
}
\examples{
# Create two correlated vectors
d1 <- rnorm(10,1)
d2 <- d1 + rnorm(10,1,0.5)

# Calculate the cosine similarity between them
cosineSimilarity(d1,d2)

}
