% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_set_annotations}
\alias{reference_set_annotations}
\title{Metadata for the whol eblood gene expression data set}
\format{
A data frame with 3264 rows and 5 variables:
\describe{
\item{Study}{Cohort or study from which the samples were originally obtained}
\item{Assay}{Type of technology used to measure gene expresison in that sample}
\item{SRS}{Predicted sepsis response signature (SRS) group for that sample}
\item{SRSq}{Predicted sepsis response signature score (SRSq) for that sample}
\item{mNN_outlier}{Indicates whether the sample is outlying. (FALSE for all samples in the reference set).}
...
}
}
\usage{
reference_set_annotations
}
\description{
A data set containing additional information for the 3,264 individuals contained in the referenc set.
}
\details{
It is formed of 5 variables gathered for 3,264 individuals

This data set is useful for visualisations and is called internally by the plotting function plotAlignedSamples().
}
\keyword{datasets}
