% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifyPatients.R
\name{stratifyPatients}
\alias{stratifyPatients}
\title{Assign samples to sepsis response signature groups}
\usage{
stratifyPatients(dat, k = 20, verbose = T)
}
\arguments{
\item{dat}{A data.frame containing n samples (rows) and 7 genes (columns). This data.frame must include all of the following columns: ENSG00000152219, ENSG00000100814, ENSG00000127334, ENSG00000131355, ENSG00000137337, ENSG00000156414, and ENSG00000115085}

\item{k}{A numeric value specifying the number of nearest neighbours used to align the input to the reference data. This parameter is used for mutual nearest neighbours alignment with mnnCorrect().}

\item{verbose}{A logical value indicating whether or not to print a step by step summary of the function's progress.}
}
\value{
A SepsisPrediction object containing the SRS and SRSq predictions for each sample, an indicator variable flagging potential outlier samples, and the predictor genes before and after mNN alignment.
}
\description{
Given a set of samples for which the seven signature genes were measured, classify
the samples into one of three sepsis reponse signature (SRS) groups and predict a quantitative
sepsis response score (SRSq) for each of them.
}
\examples{
# Load test data set
data(test_data)

# Stratify patients
predictions <- stratifyPatients(test_data)
predictions
}
