% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Whole blood gene expression measurements of 25 individuals}
\format{
A data frame with 25 rows and 7 variables:
\describe{
\item{ENSG00000152219}{ARL14EP, log-transformed gene expression level}
\item{ENSG00000100814}{CCNB1IP1, log-transformedgene expression level}
\item{ENSG00000127334}{DYRK2, log-transformed gene expression level}
\item{ENSG00000131355}{ADGRE3, log-transformed gene expression level}
\item{ENSG00000137337}{MDC1, log-transformed gene expression level}
\item{ENSG00000156414}{TDRD9, log-transformed gene expression level}
\item{ENSG00000115085}{ZAP70, log-transformed gene expression level}

...
}
}
\usage{
test_data
}
\description{
A data set containing gene expression measurements from whole blood for the 7 SRS signature genes.
}
\details{
This data set is formed of 25 patients randomly selected from the COVID-19 Blood Cell Atlas (COMBAT) study (Oxford, UK).
Gene expression was profiled using total RNA-sequencing.

The values reported in this data set were obtained after batch correction and log transformation of the raw RNA-seq counts.

The main use of this data set is to serve as a small test set with which to test the functionality of this package.
}
\keyword{datasets}
