% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Whole blood gene expression measurements for 143 samples of COVID-19 patients and healthy volunteers}
\format{
A data frame with 143 rows and 19 variables:
\describe{
\item{ENSG00000144659}{SLC25A38, log-transformed gene expression level}
\item{ENSG00000103423}{DNAJA3, log-transformed gene expression level}
\item{ENSG00000135372}{NAT10, log-transformed gene expression level}
\item{ENSG00000079134}{THOC1, log-transformed gene expression level}
\item{ENSG00000135972}{MRPS9, log-transformed gene expression level}
\item{ENSG00000087157}{PGS1, log-transformed gene expression level}
\item{ENSG00000165006}{UBAP1, log-transformed gene expression level}
\item{ENSG00000111667}{USP5, log-transformed gene expression level}
\item{ENSG00000182670}{TTC3, log-transformed gene expression level}
\item{ENSG00000097033}{SH3GLB1, log-transformed gene expression level}
\item{ENSG00000165733}{BMS1, log-transformed gene expression level}
\item{ENSG00000103264}{FBXO31, log-transformed gene expression level}
\item{ENSG00000152219}{ARL14EP, log-transformed gene expression level}
\item{ENSG00000100814}{CCNB1IP1, log-transformed gene expression level}
\item{ENSG00000127334}{DYRK2, log-transformed gene expression level}
\item{ENSG00000131355}{ADGRE3, log-transformed gene expression level}
\item{ENSG00000137337}{MDC1, log-transformed gene expression level}
\item{ENSG00000156414}{TDRD9, log-transformed gene expression level}
\item{ENSG00000115085}{ZAP70, log-transformed gene expression level}
...
}
}
\usage{
test_data
}
\description{
An example data set containing gene expression measurements from whole blood for 19 genes
}
\details{
This data set is formed of 149 samples obtained from 79 COVID-19 patients and 10 healthy volunteers from the COMBAT study (https://doi.org/10.5281/zenodo.6120249).

The values reported in this data set were obtained by log transformation of raw counts obtained from polyA-based RNA-sequencing.

The main use of this data set is to serve as a test set with which to assess the functionality of this package.
}
\keyword{datasets}
